feat(wasm): build the compiler as freestanding wasm with plain clang; drop Emscripten
Replace the Emscripten-based playground/embed build with a plain clang +
wasm-ld toolchain targeting wasm32-unknown-unknown. The compiler becomes an
ordinary linear-memory wasm module backed by a hand-rolled freestanding libc,
suitable for the in-browser playground and for embedding into a C host (e.g. a
game engine) that runs wasm the same way. (The module's linear memory is a
separate concern from the "no linear memory" rule, which governs the Lua
programs the compiler emits.)

- src/freestanding/: minimal libc — mem/str/ctype, an explicit free-list
  allocator with boundary-tag coalescing, and a vsnprintf whose %g matches
  glibc, plus the headers the compiler core needs.
- third_party/: vendored David Gay dtoa.c (strtod + correctly-rounded %.17g)
  and the Emscripten Wasm-EH setjmp/longjmp support shim, each with its
  upstream LICENSE.
- src/wasm_entry.c: replaces emscripten_entry.c; exports via the export_name
  attribute. The module imports only env.abort/env.log and exports the five
  lua2wasm_* entries plus malloc/free/memory.
- runtime/lua2wasm-wasm.mjs: host glue (string marshaling over linear memory,
  grow-safe). The playground now loads this instead of the emcc output.
- scripts/build-wasm.sh: rewritten for clang/wasm-ld (-mllvm -wasm-enable-sjlj
  + the EH tag), no emcc.
- codegen.c: compute the strpool FNV hash in uint64_t so it is independent of
  size_t width (wasm32 size_t is 32-bit); emitted output is unaffected.
- tests: test_freestanding asserts the freestanding build's WAT and assembled
  wasm are byte-identical to the native build across every e2e fixture;
  test_freestanding_fmt diffs the freestanding number formatter/parser against
  glibc over a large corpus.
- Pages CI: build with LLVM 19 (clang + lld) instead of emsdk.

+++ src/codegen.c
@@ -81,12 +81,15 @@ typedef struct {
 } StrRef;
 
 static size_t strpool_hash(const char *bytes, size_t len) {
-    size_t h = 1469598103934665603u; /* FNV-1a 64-bit offset basis */
+    /* Compute the full 64-bit FNV-1a regardless of size_t width (size_t is
+     * 32-bit on the wasm32 build) so the hash is target-independent; it only
+     * drives bucket placement, but keeping it identical avoids surprises. */
+    uint64_t h = 1469598103934665603u; /* FNV-1a 64-bit offset basis */
     for (size_t i = 0; i < len; i++) {
         h ^= (unsigned char)bytes[i];
         h *= 1099511628211u; /* FNV prime */
     }
-    return h;
+    return (size_t)h;
 }
 
 /* Brute-force fallback: find an existing run of exactly these bytes anywhere
